+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This change
 - fixed broken welcome message on `start!`
 - fixed the preparation of :kind/map - just print where possible
 - made `:format [:html]` the default
+- support for `make!` with a single value, not just a single form
 
 ## [2-alpha48] - 2023-12-04
 - making sure the base-target-directory exists before synching resources